Move `World` to separate module
