+++ core/lib/multivm/src/versions/vm_fast/mod.rs
@@ -15,3 +15,4 @@ mod tracers;
 mod utils;
 mod version;
 mod vm;
+mod world;
